fix broken react-calendar navigation
Prior to this commit, it was not possible to use the '<', '<<', '>'
and '>>' buttons on the StyledCalendar component to navigate the
months and years of the calendar on comic/ page's Ego Gala section.

I'm not sure why switching activeStartDate -> defaultActiveStartDate
worked, because the documentation doesn't clearly explain what the
difference is between a `controlled` calendar and an `uncontrolled`
calendar... but this works and doesn't seem to break so we're going with
it.